[WARP-SPEED] manage record indexes in java
(cherry picked from commit 85c71373991941a50878a6bdba9b4b0bd9675713)

+++ plugin/starburst-warp-speed/starburst-warp-speed-core/src/main/java/io/trino/plugin/warp/storage/read/StoreRowListResult.java
@@ -19,7 +19,7 @@ import java.util.Optional;
 
 public record StoreRowListResult(RecordIndexListType storeRowListType,
         int storeRowListSize,
-        Optional<Short> storeRowListStart,
+        Optional<Integer> storeRowListStart,
         int storedChunkIx)
 {
 }
